fix(example): correct the temperature object against the shipped SOMANET ESI
Verified 0x2030/0x2031 in libs/etg/tests/data/somanet-v5.6.6.xml rather than
assuming, and two of the three details were wrong in ways that would each have
stopped the motor while looking like a broken example.

The type is DINT, not INTEGER16. Device::value<T>() is type-exact, so the read
returned nothing on every real drive — and this interlock treats no reading as
too hot, so the drive would never have enabled.

The unit is 0xFD2D0000: prefix 0xFD (milli) on notation 0x2D (degree celsius), so
the object reports milli-degrees. A limit of 60 meant 0.06 °C, which trips
instantly. It is now 60000.

The object is 0x2031:01 "Drive temperature", measured near the inverter, rather
than 0x2030:01 "Core temperature" on the control board — the inverter is what a
power-stage interlock cares about. main.cc's keepFresh names the same object, and
the docs say what to check when substituting another drive's: index, type and
unit, all three silent when wrong.

+++ libs/example/example_cyclic_task.cc
@@ -99,10 +99,12 @@ void ExampleCyclicTask::execute(const CycleContext& /*ctx*/) {
   // reach into the control program.
   //
   // **No reading counts as unsafe.** An interlock that runs the motor whenever it cannot see the
-  // temperature is not an interlock. If the drive never spins, that is the first thing to check —
-  // usually keepFresh naming a different object than Config does, or the wrong data type.
+  // temperature is not an interlock. If the drive never spins, that is the first thing to check,
+  // and there are three ways to get it wrong, all silent: keepFresh naming a different object than
+  // Config does; the wrong type here (the read is type-exact, and 0x2031:01 is a DINT); or a limit
+  // in the wrong unit (it is milli-degrees, so 60 °C is 60000).
   const auto temperature =
-      drive->value<int16_t>(config_.temperatureIndex, config_.temperatureSubindex);
+      drive->value<int32_t>(config_.temperatureIndex, config_.temperatureSubindex);
   if (!temperature || *temperature > config_.temperatureLimit) {
     if (state == cia402::State::kOperationEnabled) {
       // Quick stop rather than removing the setpoint: the drive decelerates on its own
